+++ packages/block-library/src/embed/embed-bottom-sheet.native.js
@@ -86,6 +86,7 @@ const EmbedBottomSheet = ( {
 			onDismiss={ onDismiss }
 			setAttributes={ setAttributes }
 			options={ linkSettingsOptions }
+			testID="embed-edit-url-modal"
 			withBottomSheet={ withBottomSheet }
 			showIcon
 			hasPicker={ ! withBottomSheet }