+++ packages/block-library/src/template-part/index.php
@@ -55,7 +55,7 @@ function render_block_core_template_part( $attributes ) {
 			// render the corresponding file content.
 			$template_part_file_path = get_theme_file_path( '/block-template-parts/' . $attributes['slug'] . '.html' );
 			if ( 0 === validate_file( $attributes['slug'] ) && file_exists( $template_part_file_path ) ) {
-				$content = _gutenberg_inject_theme_attribute_in_content( file_get_contents( $template_part_file_path ) );
+				$content = _inject_theme_attribute_in_block_template_content( file_get_contents( $template_part_file_path ) );
 			}
 		}
 	}
@@ -99,7 +99,7 @@ function render_block_core_template_part( $attributes ) {
 	$content = $wp_embed->autoembed( $content );
 
 	if ( empty( $attributes['tagName'] ) ) {
-		$defined_areas = gutenberg_get_allowed_template_part_areas();
+		$defined_areas = get_allowed_block_template_part_areas();
 		$area_tag      = 'div';
 		foreach ( $defined_areas as $defined_area ) {
 			if ( $defined_area['area'] === $area && isset( $defined_area['area_tag'] ) ) {
@@ -122,7 +122,7 @@ function render_block_core_template_part( $attributes ) {
  */
 function build_template_part_block_variations() {
 	$variations    = array();
-	$defined_areas = gutenberg_get_allowed_template_part_areas();
+	$defined_areas = get_allowed_block_template_part_areas();
 	foreach ( $defined_areas as $area ) {
 		if ( 'uncategorized' !== $area['area'] ) {
 			$variations[] = array(