+++ packages/components/src/mobile/bottom-sheet/link-cell.native.js
@@ -12,14 +12,21 @@ import styles from './styles.scss';
 
 const { placeholderColor } = styles;
 
-export default function LinkCell( { value, onPress, showIcon = true } ) {
+export default function LinkCell( {
+	value,
+	valueMask,
+	onPress,
+	showIcon = true,
+} ) {
 	return (
 		<Cell
 			icon={ showIcon && link }
 			label={ __( 'Link to' ) }
 			// since this is not actually editable, we treat value as a placeholder
-			value={ value || __( 'Search or type URL' ) }
-			valueStyle={ !! value ? undefined : placeholderColor }
+			value={ valueMask || value || __( 'Search or type URL' ) }
+			valueStyle={
+				!! ( value || valueMask ) ? undefined : placeholderColor
+			}
 			onPress={ onPress }
 		>
 			<Icon icon={ chevronRight }></Icon>