+++ packages/edit-navigation/src/utils/index.js
@@ -1,4 +1,9 @@
 /**
+ * WordPress dependencies
+ */
+import { normalizePath } from '@wordpress/url';
+
+/**
  * The purpose of this function is to create a middleware that is responsible for preloading menu-related data.
  * It uses data that is returned from the /__experimental/menus endpoint for requests
  * to the /__experimental/menu/<menuId> endpoint, because the data is the same.
@@ -10,7 +15,7 @@
  */
 export function createMenuPreloadingMiddleware( preloadedData ) {
 	const cache = Object.keys( preloadedData ).reduce( ( result, path ) => {
-		result[ getStablePath( path ) ] = preloadedData[ path ];
+		result[ normalizePath( path ) ] = preloadedData[ path ];
 		return result;
 	}, /** @type {Record<string, any>} */ ( {} ) );
 
@@ -28,7 +33,7 @@ export function createMenuPreloadingMiddleware( preloadedData ) {
 			return next( options );
 		}
 
-		const path = getStablePath( options.path );
+		const path = normalizePath( options.path );
 		if ( ! menusDataLoaded && cache[ path ] ) {
 			menusDataLoaded = true;
 			return sendSuccessResponse( cache[ path ], parse );
@@ -83,40 +88,5 @@ function sendSuccessResponse( responseData, parse ) {
 					statusText: 'OK',
 					headers: responseData.headers,
 			  } )
-	);
-}
-
-/**
- * Given a path, returns a normalized path where equal query parameter values
- * will be treated as identical, regardless of order they appear in the original
- * text.
- *
- * @param {string} path Original path.
- *
- * @return {string} Normalized path.
- */
-export function getStablePath( path ) {
-	const splitted = path.split( '?' );
-	const query = splitted[ 1 ];
-	const base = splitted[ 0 ];
-	if ( ! query ) {
-		return base;
-	}
-
-	// 'b=1&c=2&a=5'
-	return (
-		base +
-		'?' +
-		query
-			// [ 'b=1', 'c=2', 'a=5' ]
-			.split( '&' )
-			// [ [ 'b, '1' ], [ 'c', '2' ], [ 'a', '5' ] ]
-			.map( ( entry ) => entry.split( '=' ) )
-			// [ [ 'a', '5' ], [ 'b, '1' ], [ 'c', '2' ] ]
-			.sort( ( a, b ) => a[ 0 ].localeCompare( b[ 0 ] ) )
-			// [ 'a=5', 'b=1', 'c=2' ]
-			.map( ( pair ) => pair.join( '=' ) )
-			// 'a=5&b=1&c=2'
-			.join( '&' )
 	);
 }